KVM: Reporting page poisoning value to the host
This patch enables the kernel to report the page poisoning value
to the host if Guest Free Page Hinting config is enabled. It uses
VIRTIO_BALLOON_F_PAGE_POISON feature to do so.
It is required to avoid possible guest memory corruption errors.
The issue arises when the pattern used for Page Poisoning is 0xaa while
the newly allocated page received from the hypervisor by the guest is
filled with the pattern 0x00. This will result in memory corruption errors.

The device side support for VIRTIO_BALLOON_F_PAGE_POISON is
still missing.

Signed-off-by: Nitesh Narayan Lal <nitesh@redhat.com>