KVM: Enables the kernel to isolate guest free pages
This patch enables the kernel to scan the bitmap for each zone
whose free memory threshold has met or exceed the
HINTING_THRESHOLD(16 or 64 MB if HINTING ORDER is MAX_ORDER - 1).
guest_free_page_hinting() scans the entire bitmap by
acquiring a zone lock corresponding to the pages which are
being scanned. If the page is still free and present in the
buddy it tries to isolate the page and adds it to a
dynamically allocated array.

Once this scanning process is complete and if the number of isolated
pages meets or exceeds the HINTING_THRESHOLD(16) those pages are reported
via guest_free_page_report.

For now guest_free_page_report() simply releases the pages back
to the buddy by using __free_one_page(). Once returned the scanning process
continues until the entire bitmap for the invoked zone has been scanned.
In a situation where the entire bitmap array has been scanned but the number
of isolated pages does not meets the HINTING_THRESHOLD cirteira, those pages
are released back to the buddy by acquiring a zone lock.

Signed-off-by: Nitesh Narayan Lal <nitesh@redhat.com>