KVM: Enables the kernel to isolate guest free pages
This patch enables the kernel to scan the bitmap for each zone
whose free memory count(free_mem_cnt) has met or exceeds the
HINTING_THRESHOLD(16 or 64 MB if PAGE_HINTING_MIN_ORDER is
MAX_ORDER - 1).

scan_hinting_bitmap() scans the entire bitmap by
acquiring a zone lock corresponding to the zone which is
being scanned. If the page is still free and present in the
buddy it tries to isolate the page and adds it to a
dynamically allocated array.

Once this scanning process is complete and if the number of isolated
pages meet or exceeds the HINTING_THRESHOLD(16) those pages are
reported to the host via page_hinting_report().

For now page_hinting_report() simply releases the pages back
to the buddy by using __free_one_page() and their respective bit in the
bitmap is also cleared.

Once returned the scanning process continues until the entire bitmap
for the invoked zone has been scanned. In a situation where the entire
bitmap array has been scanned but the number of isolated pages does not
meets the HINTING_THRESHOLD criteria, those pages are released back to
the buddy by acquiring a zone lock.

Signed-off-by: Nitesh Narayan Lal <nitesh@redhat.com>