KVM: Enables the kernel to report isolated pages
This patch enables the kernel to report the isolated pages
to the host via virtio balloon driver in a synchronous manner.
In order to do so, a new virtuqeue (hinting_vq) is added to the
virtio balloon driver.

Until the request is not complete no other isolation request is
generated. On completion of the request, the isolated pages are
returned back to the buddy via __free_one_page() and their
respective bits are cleared in the bitmap.

Page hinting is enabled as soon as the virtio-balloon driver is
loaded and gets disabled with its removal.

It also allows to enable/disable the page hinting feature via
from userspace by using virtio_balloon parameter page_hinting_flag.
By default, it is always enabled.

Signed-off-by: Nitesh Narayan Lal <nitesh@redhat.com>