KVM: Enables the kernel to report isolated pages
This patch adds the required callbacks which are invoked from
scan_hinting_bitmap(), i.e.,page_hinting_prepare to allocate
a space to hold the isolated page, hint_page() to add
the isolated pages into the earlier allocated space and
to report those pages if number of entries reaches/exceeds
HINTING_MEM_THRESHOLD(16), once that is done those pages are
returned back to the buddy and at the end of scanning bitmap
page_hinting_cleanup() is used to report any lingering isolated
entries and free the array at the end.

Page hinting is enabled as soon as the virtio-balloon driver is
loaded and gets disabled with its removal. It can further be
controlled by writing page_hinting_flag parameter.

Signed-off-by: Nitesh Narayan Lal <nitesh@redhat.com>