KVM: Enables the kernel to isolate guest free pages
This patch enables the kernel to scan the bitmap for each zone
whose free memory count(free_mem_cnt) has met or exceeds the
HINTING_THRESHOLD(16 or 64 MB if PAGE_HINTING_MIN_ORDER is
MAX_ORDER - 1).

scan_hinting_bitmap() scans the entire bitmap by
acquiring a zone lock corresponding to the zone which is
being scanned. It sets up the array to hold the isolated pages by
using prepare(). If the scanned page is still free and present
in the buddy it tries to isolate the page and sends it via
hint_page() to store in the array.

Once returned the scanning process continues until the entire bitmap
for the invoked zone has been scanned. After this cleanup()
is invoked to ensure that there are no pending isolated pages
remaining which are not hinted.

Signed-off-by: Nitesh Narayan Lal <nitesh@redhat.com>