mm: Hoepfully fixed the crash
Crash was probably due to the pending hint pages release.
As the local list we are maintaing will have past entries or
page references. When list read will access them and place them
in the free list, there is no real page behind. This could
lead to non-present page?

Right now we are using if check for zero bitnr value.
May be we should use list_del()? But does it removes just one
entry or the entire list?
Shouldn't we use something like list_del_entry()?

We can use __list_del_entry(), is it safe?

Signed-off-by: Nitesh Narayan Lal <nitesh@redhat.com>