mm: Adjust shuffle code to allow for future coalescing
Move the head/tail adding logic out of the shuffle code and into the
__free_one_page function since ultimately that is where it is really
needed anyway. By doing this we should be able to reduce the overhead
and can consolidate all of the list addition bits in one spot.

While changing out the code I also opted to go for a bit more thread safe
approach to getting the boolean value. This way we can avoid possible cache
line bouncing of the batched entropy between CPUs.

Signed-off-by: Alexander Duyck <alexander.h.duyck@linux.intel.com>