mm: page_reporting: core infrastructure
This patch introduces the core infrastructure for free page reporting in
virtual environments. It enables the kernel to track the free pages which
can be reported to its hypervisor so that the hypervisor could
free and reuse that memory as per its requirement.

While the pages are getting processed in the hypervisor (e.g.,
via MADV_DONTNEED), the guest must not use them, otherwise, data loss
would be possible. To avoid such a situation, these pages are
temporarily marked as MIGRATE_ISOLATE by using the existing page-isolation
APIs. The amount of pages marked temporarily from the buddy is governed
by the backend(virtio-balloon in our case). Once processed the pages
are restored via undo_isolate_page_range() for the use in VM.

To efficiently identify free pages that can to be reported to the
hypervisor, bitmaps in a coarse granularity are used. Only fairly big
chunks are reported to the hypervisor - especially, to not break up THP
in the hypervisor - "pageblock_order" on x86, and to save space. The bits
in the bitmap are an indication whether a page *might* be free, not a
guarantee. A new hook after buddy merging sets the bits.

Bitmap and its respective fields are maintained in a structure whose RCU
protected pointer is maintained on a per-zone basis. A workqueue
asynchronously processes the bitmaps, trying to isolate and report pages
that are still free.

The backend (virtio-balloon) is responsible for reporting these batched
pages to the host synchronously. Once reporting/freeing is complete,
isolated pages are returned back to the buddy.

Signed-off-by: Nitesh Narayan Lal <nitesh@redhat.com>